startedByNuxwdog(): extract common code to class
Several classes have duplicated the startedByNuxwdog subroutine.
Extract it to a utility class.
